Ben (#34)
* generation optimizations

* major generator improvements

* minor

+++ src/social_stats.py
@@ -1,6 +1,8 @@
-avarage_family_size=5
-family_strength=0.4
-avarage_work_size=50
-work_strength=0.04
+import numpy as np
+
+avarage_family_size = 5
+family_strength = 0.4
+avarage_work_size = 50
+work_strength = np.float32(0.04)
 average_amount_of_strangers = 20
-stranger_strength=0.004
+stranger_strength = 0.004
